requirements.txt
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
commandlines==0.4.1
cryptography==3.4.7
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
docutils==0.17.1
ffmpeg-python==0.2.0
filelock==3.0.12
filetype==1.0.7
future==0.18.2
gdown==3.13.0
hsh==1.1.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==4.6.1
jeepney==0.6.0
keyring==23.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.3.4
more-itertools==8.7.0
moviepy==1.0.3
networkx==2.5.1
numba==0.53.1
numpy==1.19.5
packaging==21.0
Pillow==8.1.1
pkg-resources==0.0.0
pkginfo==1.7.0
proglog==0.1.9
pycparser==2.20
Pygments==2.9.0
PyMatting==1.1.3
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
PyWavelets==1.1.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
scikit-image==0.17.2
scipy==1.5.4
SecretStorage==3.3.1
six==1.16.0
tifffile==2020.9.3
torch==1.8.1
torchvision==0.9.1
tqdm==4.61.0
twine==3.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
waitress==2.0.0
webencodings==0.5.1
zipp==3.5.0
